refactor: Rename temp_navbar to navbar
