+++ src/pages/profile/Profile.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { MainContainer } from "@shared/ui/mainContainer";
-import { Navbar } from "@/widgets/temp_navbar";
+import { Navbar } from "@/widgets/navbar";
 import { AddShelter } from "@widgets/addShelter/index.ts";
 import * as petModel from "@entities/pet/index.ts";
 export const Profile: React.FC = () => {
