+++ src/widgets/navbar/ui/navbar.module.scss
